Lab5 done without displaying an image of an object that is being edited

+++ WEB_253504_Novikov.UI/Services/VehicleService/ApiVehicleService.cs
@@ -2,6 +2,7 @@
 using System.Text.Json;
 using WEB_253504_Novikov.Domain.Entities;
 using WEB_253504_Novikov.Domain.Models;
+using WEB_253504_Novikov.UI.Services.FileService;
 
 namespace WEB_253504_Novikov.UI.Services.VehicleService
 {
@@ -11,11 +12,13 @@ namespace WEB_253504_Novikov.UI.Services.VehicleService
         private readonly string _pageSize;
         private readonly JsonSerializerOptions _serializerOptions;
         private readonly ILogger<ApiVehicleService> _logger;
+        private readonly IFileService _fileService;
 
         public ApiVehicleService(
             HttpClient client,
             IConfiguration configuration,
-            ILogger<ApiVehicleService> logger)
+            ILogger<ApiVehicleService> logger,
+            IFileService fileService)
         {
             _httpClient = client;
             _pageSize = configuration.GetSection("ItemsPerPage").Value;
@@ -24,27 +27,59 @@ namespace WEB_253504_Novikov.UI.Services.VehicleService
                 PropertyNamingPolicy = JsonNamingPolicy.CamelCase
             };
             _logger = logger;
+            _fileService = fileService;
         }
+
+        //POST Vehicle
         public async Task<ResponseData<Vehicle>> CreateProductAsync(Vehicle product, IFormFile? formFile)
         {
+            product.ImagePath = "https://localhost:7002/Images/no-image.png";
+
+            if(formFile != null)
+            {
+                var imageUrl = await _fileService.SaveFileAsync(formFile);
+                if (!string.IsNullOrEmpty(imageUrl))
+                    product.ImagePath = imageUrl;
+            }
+
             var uri = new Uri(_httpClient.BaseAddress.AbsoluteUri + "Vehicles");
 
-            var response = await _httpClient.PostAsJsonAsync(
+            var multipartContent = new MultipartFormDataContent
+            {
+                { new StringContent(product.Name), "Name" },
+                { new StringContent(product.Description), "Description" },
+                { new StringContent(product.TypeId.ToString()), "TypeId" },
+                { new StringContent(product.Cost.ToString()), "Cost" },
+                { new StringContent(product.ImagePath), "ImagePath" }
+            };
+
+
+            var response = await _httpClient.PostAsync(
                 uri,
-                product,
-                _serializerOptions);
+                multipartContent);
             if (response.IsSuccessStatusCode)
             {
                 var data = await response.Content.ReadFromJsonAsync<ResponseData<Vehicle>>(_serializerOptions);
                 return data; // Vehicle;
             }
-            _logger.LogError($"-----> object not created. Error:{ response.StatusCode.ToString()}");
+            _logger.LogError($"-----> object not created. Error:{response.StatusCode.ToString()}");
             return ResponseData<Vehicle>.Error($"Объект не добавлен. Error:{response.StatusCode.ToString()}");
         }
 
+        //DELETE Vehicle
         public async Task<ResponseData<Vehicle>> DeleteProductAsync(int id)
         {
             var uri = new Uri(_httpClient.BaseAddress.AbsoluteUri + "Vehicles/" + id.ToString());
+
+            var productResponse = await _httpClient.GetAsync(uri);
+            if (!productResponse.IsSuccessStatusCode)
+            {
+                _logger.LogError($"-----> object not deleted. Error:{productResponse.StatusCode.ToString()}");
+                return ResponseData<Vehicle>.Error($"Объект не удален. Error:{productResponse.StatusCode.ToString()}");
+            }
+            var product = await productResponse.Content.ReadFromJsonAsync<ResponseData<Vehicle>>(_serializerOptions);
+
+            await _fileService.DeleteFileAsync(product.Data.ImagePath);
 
             var response = await _httpClient.DeleteAsync(uri);
             if (response.IsSuccessStatusCode)
@@ -56,6 +91,7 @@ namespace WEB_253504_Novikov.UI.Services.VehicleService
             return ResponseData<Vehicle>.Error($"Объект не удален. Error:{response.StatusCode.ToString()}");
         }
 
+        //GET Vehicle
         public async Task<ResponseData<Vehicle>> GetProductByIdAsync(int id)
         {
             var uri = new Uri(_httpClient.BaseAddress.AbsoluteUri + "Vehicles/" + id.ToString());
@@ -70,6 +106,7 @@ namespace WEB_253504_Novikov.UI.Services.VehicleService
             return ResponseData<Vehicle>.Error($"Объект не получен. Error:{response.StatusCode.ToString()}");
         }
 
+        //GET Vehicle List
         public async Task<ResponseData<ListModel<Vehicle>>> GetProductListAsync(string? categoryNormalizedName = null, int pageNo = 1)
         {
             var urlString = new StringBuilder($"{_httpClient.BaseAddress.AbsoluteUri}Vehicles");
@@ -106,9 +143,18 @@ namespace WEB_253504_Novikov.UI.Services.VehicleService
             return ResponseData<ListModel<Vehicle>>.Error($"Данные не получены от сервера. Error{response.StatusCode.ToString()}");
         }
 
+        //UPDATE
         public async Task<ResponseData<Vehicle>> UpdateProductAsync(int id, Vehicle product, IFormFile? formFile)
         {
             var uri = new Uri(_httpClient.BaseAddress.AbsoluteUri + "Vehicles/" + id.ToString());
+
+            if (formFile != null)
+            {
+                await _fileService.DeleteFileAsync(product.ImagePath);
+                var imageUrl = await _fileService.SaveFileAsync(formFile);
+                if (!string.IsNullOrEmpty(imageUrl))
+                    product.ImagePath = imageUrl;
+            }
 
             var response = await _httpClient.PutAsJsonAsync(
                 uri,
